% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icqr.R
\name{icqr}
\alias{icqr}
\title{icqr}
\usage{
icqr(
  data,
  u,
  attempts = 10,
  method = "Brent",
  lower = -5,
  upper = 5,
  deltaName = c("delta"),
  timeName = c("y"),
  covNames = c("x"),
  IVNames = c("w"),
  Nbootstrap = NA,
  bootstrapEpsilon = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{dataset as dataframe}

\item{attempts}{attempts to estimate beta, to avoid local minima}

\item{method}{estimation method, see optim function in stats package.}

\item{lower}{1 dimension: (lower,upper) = interval where beta is searched (see optim function in stats package.),more dimensions: interval for uniform random starting point for beta, in each component}

\item{upper}{see "lower" parameter}

\item{deltaName}{variable name of the censor variable: 0 censored, 1 not censored. It is passed as list of string}

\item{timeName}{variable name of the time variable.  It is passed as list of string}

\item{covNames}{variable names of the covriate. It is passed as list of string}

\item{IVNames}{variable names for the Instrumental variable, it has to include exogenuos variable name.  It is passed as list of string}

\item{Nbootstrap}{number of bootstrap resampling. NA in case of not boostrap}

\item{bootstrapEpsilon}{when bootstrap, reaserch the minimum in (beta-epsilon, beta + epsilon), vector interval}

\item{verbose}{print during estimation}
}
\value{
estimated beta vector parameter
}
\description{
This function returns the coefficient vector beta
in the model T = exp(z beta(u)) instrumental censored quantile regression.
}
